feat: добавлен запрос для получения пользователя периода
