--- client/gql/dcis/queries/user_divisions.graphql
@@ -1,7 +1,0 @@
-#import "../fragments/division_model_fields.graphql"
-
-query UserDivisions($userId: ID, $projectId: ID) {
-  userDivisions (userId: $userId, projectId: $projectId) {
-    ...DivisionModelFields
-  }
-}
