+++ client/gql/dcis/queries/user_group_privileges.graphql
@@ -1,0 +1,7 @@
+#import "../fragments/privilege_fields.graphql"
+
+query UserGroupPrivileges($userId: ID, $periodGroupId: ID!) {
+  userGroupPrivileges(userId: $userId, periodGroupId: $periodGroupId) {
+    ...PrivilegesFields
+  }
+}
