+++ client/gql/dcis/queries/user_individual_privileges.graphql
@@ -1,0 +1,7 @@
+#import "../fragments/privilege_fields.graphql"
+
+query UserIndividualPrivileges($userId: ID, $periodId: ID!) {
+  userIndividualPrivileges(userId: $userId, periodId: $periodId) {
+    ...PrivilegesFields
+  }
+}
