--- client/gql/dcis/queries/user_period_privileges.graphql
@@ -1,7 +1,0 @@
-#import "../fragments/privilege_fields.graphql"
-
-query UserIndividualPrivileges($periodId: ID!, $userId: ID = null) {
-  userIndividualPrivileges(periodId: $periodId, userId: $userId) {
-    ...PrivilegesFields
-  }
-}
