+++ server/apps/dcis/schema/queries/__init__.py
@@ -1,4 +1,3 @@
 from .document_queries import DocumentQueries
-from .privilege_queries import PrivilegeQueries
 from .project_queries import ProjectQueries
-from .sheet_queries import SheetQueries
+from .period_queries import PeriodQueries
